# batch-console-api/src/main/java/com/example/batch/console/web/ConsoleWorkflowRunController.java
package com.example.batch.console.web;

import com.example.batch.common.dto.CommonResponse;
import com.example.batch.console.application.ConsoleOrchestratorProxyService;
import com.example.batch.console.service.ConsoleResponseFactory;
import java.util.Map;
import lombok.RequiredArgsConstructor;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

@RestController
@Validated
@RequestMapping("/api/console/workflow-runs")
@PreAuthorize("hasAuthority('ROLE_ADMIN')")
@RequiredArgsConstructor
public class ConsoleWorkflowRunController {

    private final ConsoleOrchestratorProxyService orchestratorProxyService;
    private final ConsoleResponseFactory responseFactory;

    @PostMapping("/{id}/cancel")
    public CommonResponse<Map<String, Object>> cancel(@PathVariable Long id,
                                                       @RequestParam("tenantId") String tenantId) {
        return responseFactory.success(orchestratorProxyService.workflowRunAction(id, tenantId, "cancel"));
    }

    @PostMapping("/{id}/terminate")
    public CommonResponse<Map<String, Object>> terminate(@PathVariable Long id,
                                                          @RequestParam("tenantId") String tenantId) {
        return responseFactory.success(orchestratorProxyService.workflowRunAction(id, tenantId, "terminate"));
    }

    @PostMapping("/{id}/skip-node")
    public CommonResponse<Map<String, Object>> skipNode(@PathVariable Long id,
                                                         @RequestParam("tenantId") String tenantId,
                                                         @RequestParam("nodeCode") String nodeCode) {
        return responseFactory.success(orchestratorProxyService.workflowRunSkipNode(id, tenantId, nodeCode));
    }
}