# docs/api/console-api.openapi.yaml
openapi: 3.1.0
info:
  title: Batch Platform Console API
  version: 0.1.0
  description: >
    Frontend-facing contract for the batch console.
    This file is the machine-readable source of truth for UI integration,
    code generation, and mock server stubs.
    Whenever the API surface changes, update this file and console-api-protocol.md together.
servers:
  - url: http://localhost:8080
    description: Local console-api
...
security:
  - bearerAuth: []
  - consoleToken: []
paths:
...
  /api/console/scheduler/snapshot:
    get:
      tags: [Scheduler]
      summary: Get live scheduler snapshot
      operationId: getSchedulerSnapshot
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
      responses:
        '200':
          description: Snapshot body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleSchedulerSnapshotResponse'
  /api/console/scheduler/snapshot/history:
    get:
      tags: [Scheduler]
      summary: Get scheduler snapshot history
      operationId: getSchedulerSnapshotHistory
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: Snapshot history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleSchedulerSnapshotHistoryList'
...
  /api/console/triggers:
    get:
      tags: [Triggers]
      summary: List registered Quartz triggers
      operationId: listTriggers
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trigger list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseObject'
  /api/console/triggers/{jobCode}/register:
    post:
      tags: [Triggers]
      summary: Register a Quartz trigger for a job
      operationId: registerTrigger
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: jobCode
          in: path
          required: true
          schema:
            type: string
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/triggers/{jobCode}/unregister:
    post:
      tags: [Triggers]
      summary: Unregister a Quartz trigger for a job
      operationId: unregisterTrigger
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: jobCode
          in: path
          required: true
          schema:
            type: string
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/triggers/{jobCode}/pause:
    post:
      tags: [Triggers]
      summary: Pause a Quartz trigger for a job
      operationId: pauseTrigger
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: jobCode
          in: path
          required: true
          schema:
            type: string
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/triggers/{jobCode}/resume:
    post:
      tags: [Triggers]
      summary: Resume a paused Quartz trigger for a job
      operationId: resumeTrigger
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: jobCode
          in: path
          required: true
          schema:
            type: string
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'

  # ─── P0: Enum metadata ────────────────────────────────────────────────────────
...
  /api/console/queues:
    get:
      tags: [Queues]
      summary: List resource queues
      operationId: listResourceQueues
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
        - name: queueCode
          in: query
          required: false
          schema:
            type: string
        - name: queueType
          in: query
          required: false
          schema:
            type: string
        - name: enabled
          in: query
          required: false
          schema:
            type: boolean
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Resource queue list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseObject'
    post:
      tags: [Queues]
      summary: Create resource queue
      operationId: createResourceQueue
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Created resource queue id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseLong'
  /api/console/queues/{id}:
    put:
      tags: [Queues]
      summary: Update resource queue
      operationId: updateResourceQueue
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/queues/{id}/toggle:
    post:
      tags: [Queues]
      summary: Enable or disable resource queue
      operationId: toggleResourceQueue
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
        - name: enabled
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'

  # ─── P1: Batch Windows CRUD ───────────────────────────────────────────────────
...
  /api/console/workflow-runs/{id}/cancel:
    post:
      tags: [WorkflowRuns]
      summary: Cancel a workflow run
      operationId: cancelWorkflowRun
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/workflow-runs/{id}/terminate:
    post:
      tags: [WorkflowRuns]
      summary: Terminate a workflow run
      operationId: terminateWorkflowRun
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/workflow-runs/{id}/skip-node:
    post:
      tags: [WorkflowRuns]
      summary: Skip a node in a workflow run
      operationId: skipWorkflowRunNode
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
        - name: nodeCode
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'

  # ─── P2: Dashboard ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    consoleToken:
      type: apiKey
      in: header
      name: X-Console-Token
  parameters:
...
    TenantIdQuery:
      name: tenantId
      in: query
      required: false
      schema:
        type: string
...
    PageNoQuery:
      name: pageNo
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 20
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  schemas:
    CommonMeta:
      type: object
      properties:
        requestId:
          type: string
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [requestId, traceId, timestamp]
    CommonResponseBase:
      type: object
      properties:
        code:
          type: string
          description: Application-level status code.
        message:
          type: string
          description: Human-readable status message.
        meta:
          $ref: '#/components/schemas/CommonMeta'
      required: [code, message, meta]
...
    CommonResponseString:
      allOf:
        - $ref: '#/components/schemas/CommonResponseBase'
        - type: object
          properties:
            data:
              type: string
    CommonResponseLong:
      allOf:
        - $ref: '#/components/schemas/CommonResponseBase'
        - type: object
          properties:
            data:
              type: integer
              format: int64
...
    CommonResponseObject:
      allOf:
        - $ref: '#/components/schemas/CommonResponseBase'
        - type: object
          properties:
            data:
              nullable: true
...
    CommonResponseConsoleSchedulerSnapshotResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ConsoleSchedulerSnapshotResponse'
    CommonResponseConsoleSchedulerSnapshotHistoryList:
      allOf:
        - $ref: '#/components/schemas/CommonResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ConsoleSchedulerSnapshotHistoryResponse'
...
    ConsoleSchedulerSnapshotResponse:
      type: object
      properties:
        generatedAt:
          type: string
          format: date-time
        tenantId:
          type: string
        policies:
          type: array
          items:
            $ref: '#/components/schemas/ConsoleSchedulerPolicySnapshot'
        queues:
          type: array
          items:
            $ref: '#/components/schemas/ConsoleSchedulerQueueSnapshot'
        workers:
          type: array
          items:
            $ref: '#/components/schemas/ConsoleSchedulerWorkerLoadSnapshot'
      required: [generatedAt, tenantId, policies, queues, workers]
    ConsoleSchedulerPolicySnapshot:
      type: object
      properties:
        policyCode:
          type: string
        fairShareGroup:
          type: string
        fairShareWeight:
          type: integer
          format: int32
        maxRunningJobsPerTenant:
          type: integer
          format: int32
        burstLimit:
          type: integer
          format: int32
        partitionBurstLimit:
          type: integer
          format: int32
        quotaResetPolicy:
          type: string
        quotaBurstPeakBorrowed:
          type: integer
          format: int32
        quotaBurstRemaining:
          type: integer
          format: int32
        quotaResetWindowStartedAt:
          type: string
          format: date-time
        quotaResetWindowExpiresAt:
          type: string
          format: date-time
        groupSharedMaxRunningJobs:
          type: integer
          format: int32
        activeJobs:
          type: integer
          format: int64
        activePartitions:
          type: integer
          format: int64
        groupActiveJobs:
          type: integer
          format: int64
        effectiveTenantJobCap:
          type: integer
          format: int32
        effectiveTenantPartitionCap:
          type: integer
          format: int32
      required: [policyCode, fairShareGroup, fairShareWeight, maxRunningJobsPerTenant, burstLimit, partitionBurstLimit, quotaResetPolicy, quotaBurstPeakBorrowed, quotaBurstRemaining, quotaResetWindowStartedAt, quotaResetWindowExpiresAt, groupSharedMaxRunningJobs, activeJobs, activePartitions, groupActiveJobs, effectiveTenantJobCap, effectiveTenantPartitionCap]
    ConsoleSchedulerQueueSnapshot:
      type: object
      properties:
        queueCode:
          type: string
        fairShareGroup:
          type: string
        fairShareWeight:
          type: integer
          format: int32
        maxRunningJobs:
          type: integer
          format: int32
        burstLimit:
          type: integer
          format: int32
        effectiveMaxRunningJobs:
          type: integer
          format: int32
        quotaResetPolicy:
          type: string
        quotaBurstPeakBorrowed:
          type: integer
          format: int32
        quotaBurstRemaining:
          type: integer
          format: int32
        quotaResetWindowStartedAt:
          type: string
          format: date-time
        quotaResetWindowExpiresAt:
          type: string
          format: date-time
        groupSharedMaxRunningJobs:
          type: integer
          format: int32
        activeJobs:
          type: integer
          format: int64
      required: [queueCode, fairShareGroup, fairShareWeight, maxRunningJobs, burstLimit, effectiveMaxRunningJobs, quotaResetPolicy, quotaBurstPeakBorrowed, quotaBurstRemaining, quotaResetWindowStartedAt, quotaResetWindowExpiresAt, groupSharedMaxRunningJobs, activeJobs]
    ConsoleSchedulerWorkerLoadSnapshot:
      type: object
      properties:
        workerCode:
          type: string
        workerGroup:
          type: string
        currentLoad:
          type: integer
          format: int32
        heartbeatAt:
          type: string
          format: date-time
        status:
          type: string
      required: [workerCode, workerGroup, currentLoad, heartbeatAt, status]
    ConsoleSchedulerSnapshotHistoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tenantId:
          type: string
        snapshotAt:
          type: string
          format: date-time
        fairShareGroup:
          type: string
        policyCode:
          type: string
        activeJobs:
          type: integer
          format: int32
        activePartitions:
          type: integer
          format: int32
        maxJobsBase:
          type: integer
          format: int32
        burstLimit:
          type: integer
          format: int32
        effectiveJobCap:
          type: integer
          format: int32
        groupActiveJobs:
          type: integer
          format: int32
        groupMaxJobs:
          type: integer
          format: int32
        quotaResetPolicy:
          type: string
        onlineWorkers:
          type: integer
          format: int32
        detailJson:
          type: string
      required: [id, tenantId, snapshotAt, fairShareGroup, policyCode, activeJobs, activePartitions, maxJobsBase, burstLimit, effectiveJobCap, groupActiveJobs, groupMaxJobs, quotaResetPolicy, onlineWorkers, detailJson]
...